+++ Frontned/shopping-app/src/components/Items.tsx
@@ -74,9 +74,9 @@ function Items() {
 					{filteredItems ? filteredItems.map((item: IItem) => {
 						return (
 							<Col sm={4} py={2}>
-								<Card className="item-card d-flex align-items-center justify-content-center" style={{ width: '18rem', height: '50rem' }}>
-									<Card.Header> <strong>{item.name}</strong> </Card.Header>
-									<Card.Img variant="top" src={item.image} />
+								<Card className="item-card d-flex align-items-center justify-content-center" style={{ width: '18rem', height: '30rem' }}>
+									<Card.Header className="item-title"> <strong>{item.name}</strong> </Card.Header>
+									<Card.Img className="item-img" height={512} variant="top" src={item.image} alt={item.name} />
 									<Card.Body>
 										<Card.Subtitle>{item.description}</Card.Subtitle>
 									</Card.Body>
