--- Frontned/shopping-app/src/components/css/footer.scss
@@ -1,4 +1,0 @@
-.footer-wrapper {
-	bottom: 0;
-	width: 100%;
-}
