+++ Frontned/shopping-app/src/index.scss
@@ -1,0 +1,183 @@
+@import url("https://fonts.googleapis.com/css2?family=Nunito+Sans:wght@700&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Cedarville+Cursive&display=swap");
+
+.title {
+  padding-top: 2%;
+}
+
+.title-line {
+  height: 15px;
+  opacity: 1;
+  border-width: 3px;
+  border-style: solid;
+}
+
+.order-text {
+  color: rgb(117, 117, 117);
+}
+
+.item-img {
+  border: solid rgb(207, 154, 40);
+  height: auto;
+  object-fit: cover;
+  max-height: 256px;
+}
+
+.rm-btn {
+  position: absolute;
+}
+
+.item-title {
+  font-family: "Nunito Sans", sans-serif;
+}
+
+.footer-wrapper {
+  bottom: 0;
+  width: 100%;
+}
+
+.title {
+  padding-top: 2%;
+}
+
+.title-line {
+  height: 15px;
+  opacity: 1;
+  border-width: 3px;
+  border-style: solid;
+}
+
+.item-card {
+  margin: 5%;
+}
+
+.tags {
+  list-style: none;
+  margin: 0;
+  overflow: hidden;
+  padding: 0;
+}
+
+.tags li {
+  float: left;
+}
+
+.tag {
+  border-radius: 3px 0 0 3px;
+  color: #999;
+  display: inline-block;
+  height: 26px;
+  line-height: 26px;
+  padding: 0 20px 0 23px;
+  position: relative;
+  margin: 0 10px 10px 0;
+  text-decoration: none;
+  transition: color 0.2s;
+  -webkit-transition: color 0.2s;
+}
+
+.tag::before {
+  border-radius: 10px;
+  box-shadow: inset 0 1px rgba(0, 0, 0, 0.25);
+  content: "";
+  height: 6px;
+  left: 10px;
+  position: absolute;
+  width: 6px;
+  top: 10px;
+}
+
+.tag::after {
+  border-bottom: 13px solid transparent;
+  border-top: 13px solid transparent;
+  content: "";
+  position: absolute;
+  right: 0;
+  top: 0;
+}
+
+.tag:hover {
+  background-color: crimson;
+  color: white;
+}
+
+.tag:hover::after {
+  border-left-color: crimson;
+}
+
+.checkout-wrapper {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  padding-top: 25vh;
+}
+
+.cursive-text {
+  font-family: "Cedarville Cursive", cursive;
+}
+
+.not-found-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding-top: 25vh;
+}
+
+.fab-item {
+  box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23);
+  border-radius: 50%;
+  border-style: none;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 56px;
+  height: 56px;
+  position: fixed;
+  right: 1%;
+  bottom: 1%;
+  -webkit-transition: transform 0.1s ease-out, height 100ms ease, opacity 100ms ease;
+  transition: transform 0.1s ease-out, height 100ms ease, opacity 100ms ease;
+  text-decoration: none;
+}
+
+.fab-item:active,
+.fab-item:focus,
+.fab-item:hover {
+  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.19), 0 6px 6px rgba(0, 0, 0, 0.23);
+  transition: box-shadow 0.2s ease;
+  outline: none;
+}
+
+.fab-item:not(:last-child) {
+  width: 40px;
+  height: 0px;
+  margin: 0px auto 0;
+  opacity: 0;
+  -webkit-transform: translateY(50px);
+  -ms-transform: translateY(50px);
+  transform: translateY(50px);
+}
+
+.login-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding-top: 25vh;
+}
+
+.edit-cards {
+  padding-left: 7vw;
+}
+
+.register-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+padding-top: 25vh;
+}
+
+.carousel-img {
+	height: auto;
+	object-fit: cover;
+	max-height: 512px;
+}
