--- Frontned/shopping-app/src/pages/cart/cart.scss
@@ -1,23 +1,0 @@
-.title {
-    padding-top: 2%;
-}
-
-.title-line {
-    height: 15px;
-    opacity: 1;
-    border-width: 3px;
-    border-style: solid;
-}
-
-.order-text {
-	color: rgb(117, 117, 117);
-}
-
-.item-img {
-    border: solid rgb(207, 154, 40);
-}
-
-.rm-btn {
-    position: absolute;
-}
-
