+++ Frontned/shopping-app/src/pages/catalog.tsx
@@ -1,0 +1,15 @@
+import Bar from "../components/Bar";
+import Footer from "../components/Footer";
+import Items from "../components/Items";
+
+function Catalog() {
+	return (
+		<>
+			<Bar />
+			<Items />
+			<Footer />
+		</>
+	)
+}
+
+export default Catalog;
