--- Frontned/shopping-app/src/pages/catalog/catalog.scss
@@ -1,68 +1,0 @@
-.title {
-  padding-top: 2%;
-}
-
-.title-line {
-  height: 15px;
-  opacity: 1;
-  border-width: 3px;
-  border-style: solid;
-}
-
-.item-card {
-  margin: 5%;
-}
-
-.tags {
-  list-style: none;
-  margin: 0;
-  overflow: hidden;
-  padding: 0;
-}
-
-.tags li {
-  float: left;
-}
-
-.tag {
-  border-radius: 3px 0 0 3px;
-  color: #999;
-  display: inline-block;
-  height: 26px;
-  line-height: 26px;
-  padding: 0 20px 0 23px;
-  position: relative;
-  margin: 0 10px 10px 0;
-  text-decoration: none;
-  transition: color 0.2s;
-  -webkit-transition: color 0.2s;
-}
-
-.tag::before {
-  border-radius: 10px;
-  box-shadow: inset 0 1px rgba(0, 0, 0, 0.25);
-  content: "";
-  height: 6px;
-  left: 10px;
-  position: absolute;
-  width: 6px;
-  top: 10px;
-}
-
-.tag::after {
-  border-bottom: 13px solid transparent;
-  border-top: 13px solid transparent;
-  content: "";
-  position: absolute;
-  right: 0;
-  top: 0;
-}
-
-.tag:hover {
-  background-color: crimson;
-  color: white;
-}
-
-.tag:hover::after {
-  border-left-color: crimson;
-}
