--- Frontned/shopping-app/src/pages/catalog/catalog.tsx
@@ -1,17 +1,0 @@
-import Bar from "../../components/Bar";
-import Footer from "../../components/Footer";
-import Items from "../../components/Items";
-
-import './catalog.scss';
-
-function Catalog() {
-	return (
-		<>
-			<Bar />
-			<Items />
-			<Footer />
-		</>
-	)
-}
-
-export default Catalog;
