--- Frontned/shopping-app/src/pages/checkout/checkout.scss
@@ -1,12 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Cedarville+Cursive&display=swap');
-
-.checkout-wrapper {
-	display: flex;
-    align-items: center;
-    flex-direction: column;
-	padding-top: 25vh;
-}
-
-.cursive-text {
-    font-family: 'Cedarville Cursive', cursive;
-}
