--- Frontned/shopping-app/src/pages/home/home.scss
@@ -1,10 +1,0 @@
-.title {
-    padding-top: 2%;
-}
-
-.title-line {
-    height: 15px;
-    opacity: 1;
-    border-width: 3px;
-    border-style: solid;
-}
