--- Frontned/shopping-app/src/pages/login/login.scss
@@ -1,6 +1,0 @@
-.login-wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-	padding-top: 25vh;
-}
