--- Frontned/shopping-app/src/pages/not-found/notFound.scss
@@ -1,10 +1,0 @@
-.not-found-wrapper {
-	display: flex;
-    flex-direction: column;
-    align-items: center;
-	padding-top: 25vh;
-}
-
-.text-wrap-center {
-    text-align: center;
-}
