--- Frontned/shopping-app/src/pages/profile/profile.scss
@@ -1,14 +1,0 @@
-.title {
-    padding-top: 2%;
-}
-
-.title-line {
-    height: 15px;
-    opacity: 1;
-    border-width: 3px;
-    border-style: solid;
-}
-
-.edit-cards {
-	padding-left: 7vw;
-}
