--- Frontned/shopping-app/src/pages/register/register.scss
@@ -1,6 +1,0 @@
-.register-wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-	padding-top: 25vh;
-}
