--- Frontned/shopping-app/src/theme/theme-btn.scss
@@ -1,34 +1,0 @@
-.fab-item {
-	box-shadow: 0 3px 6px rgba(0,0,0,0.16), 0 3px 6px rgba(0,0,0,0.23);
-	border-radius: 50%;
-	border-style: none;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	width: 56px;
-	height: 56px;
-	position: fixed;
-	right: 1%;
-	bottom: 1%;
-	-webkit-transition: transform .1s ease-out, height 100ms ease, opacity 100ms ease;
-			transition: transform .1s ease-out, height 100ms ease, opacity 100ms ease;  
-	text-decoration: none;
-  }
-  
-  .fab-item:active, 
-  .fab-item:focus, 
-  .fab-item:hover {
-	box-shadow: 0 10px 20px rgba(0,0,0,0.19), 0 6px 6px rgba(0,0,0,0.23);
-	transition: box-shadow .2s ease;
-	outline: none;
-  }
-  
-  .fab-item:not(:last-child) {
-	width: 40px;
-	height: 0px;
-	margin: 0px auto 0;
-	opacity: 0;
-	-webkit-transform: translateY(50px);
-	-ms-transform: translateY(50px);
-	transform: translateY(50px);
-  }
